fix(temporal-infra/shim): batch 6 — V8 method-signature mismatches
V8's js-temporal-objects.cc is now successfully compiling against the
shim past all the structural include / circular-dep issues. Run 12
hit 10 unique remaining method-shape mismatches:

1. RelativeTo: shim used `plain_date`/`zoned_date_time` field names;
   upstream uses `date`/`zoned`. V8 uses designated initializers
   `{.date = ..., .zoned = ...}`. Drop the Kind enum and route via
   nullness of the two pointers.

2. PartialDuration: shim had `optional<double>` for all fields;
   upstream's diplomat surface uses `optional<int64_t>` for years/
   months/weeks/days/hours/minutes/seconds/milliseconds and
   `optional<double>` only for microseconds/nanoseconds. V8's
   GetSingleDurationFieldInteger returns `optional<long long>`.

3. ParsedDateTime: add `from_utf16(u16string_view)` static factory.

4. PlainYearMonth, PlainMonthDay: add `from_partial(PartialDate,
   optional<ArithmeticOverflow>)` static factories. Stub bodies.

5. ZonedDateTime: add `from_partial_with_provider(...)` static
   factory. 5 args including PartialZonedDateTime + optional disamb +
   optional offset_option + optional overflow + Provider. Stub body.

6. PlainDate: add `to_plain_year_month()` and `to_plain_month_day()`.
   Stub bodies.

7. PlainDate::to_zoned_date_time_with_provider: was templated taking
   `(const TZ&, const Provider&)`; upstream is non-templated taking
   `(TimeZone, const PlainTime*, const Provider&)`. Replace.

8. PlainDateTime::to_zoned_date_time_with_provider: was templated
   taking `(const TZ&, const Provider&)`; upstream is non-templated
   taking `(TimeZone, Disambiguation, const Provider&)`. Replace.

PlainDate.hpp now needs PlainTime + ZonedDateTime forward decls (in
addition to existing fwds).

Down from 396 → 20 → expected near-zero in this batch.

+++ packages/temporal-infra/include/temporal_rs/PartialDuration.hpp
@@ -4,19 +4,23 @@
 #ifndef TEMPORAL_RS_COMPAT_PARTIALDURATION_HPP_
 #define TEMPORAL_RS_COMPAT_PARTIALDURATION_HPP_
 
+#include <cstdint>
 #include <optional>
 
 namespace temporal_rs {
 
 struct PartialDuration {
-  std::optional<double> years;
-  std::optional<double> months;
-  std::optional<double> weeks;
-  std::optional<double> days;
-  std::optional<double> hours;
-  std::optional<double> minutes;
-  std::optional<double> seconds;
-  std::optional<double> milliseconds;
+  // Match upstream's diplomat surface: integer fields for time units
+  // that fit exactly in int64, double only for sub-millisecond
+  // precision where IEEE-754 fractional bits matter.
+  std::optional<int64_t> years;
+  std::optional<int64_t> months;
+  std::optional<int64_t> weeks;
+  std::optional<int64_t> days;
+  std::optional<int64_t> hours;
+  std::optional<int64_t> minutes;
+  std::optional<int64_t> seconds;
+  std::optional<int64_t> milliseconds;
   std::optional<double> microseconds;
   std::optional<double> nanoseconds;
